update keyboard replay, add delete message after used in add_mode and test_mode

+++ handler/handlers_comand.py
@@ -9,42 +9,48 @@ from DB import db2
 async def start(message: types.Message):
     """Стартовая функция"""
 
-    button = ["Добавить слово ->", "Добавить фразу ->", "Пройти тест: слова ->",
-              "Пройти тест: фразы ->", "Повторение ->", "Настройки ->",
-              "Яблоко", "Apple"]
-    keyboard_start = Keyboard(button)
+    if message.from_user.id == 26750009 or message.from_user.id == 1882554481:
+        pass
 
-    await bot.send_message(message.from_user.id, f"{message_handlers.handlers_dict['start']}",
-                           reply_markup=keyboard_start.create_keyboadr())
-    await bot.delete_message(message.chat.id, message.message_id)
+    else:
 
-    create_table = db2.DB(message.from_user.id)
-    create_table.create_table()
+        button = ["Добавить слово ->", "Добавить фразу ->", "Пройти тест: слова ->",
+                  "Пройти тест: фразы ->", "Повторение слова ->", "Настройки ->"]
+        keyboard_start = Keyboard(button)
 
-    param_questions = "param_questions"
+        await bot.send_message(message.from_user.id, f"{message_handlers.handlers_dict['start']}",
+                               reply_markup=keyboard_start.create_keyboadr())
+        await bot.delete_message(message.chat.id, message.message_id)
 
-    try:
-        if int(create_table.select_data(param_questions)[0][0]) == 10:
-            pass
+        create_table = db2.DB(message.from_user.id)
+        create_table.create_table()
 
-    except Exception as ex:
+        param_questions = "param_questions"
 
-        status = 1
-        butt_dict = {
-            "1": "За посл неделю ✅",
-            "2": "За все время ",
-            "3": "рус --> англ ✅",
-            "4": "англ --> рус "
-        }
+        try:
+            if int(create_table.select_data(param_questions)[0][0]) == 10:
+                pass
 
-        butt_dict_upd = {
-            "1": "За посл неделю ",
-            "2": "За все время ",
-            "3": "рус --> англ ",
-            "4": "англ --> рус "
-        }
+        except Exception as ex:
 
-        create_table.insert_settings(3, 65, status, butt_dict, butt_dict_upd)
+            defalt_day = 7
+
+            status = 1
+            butt_dict = {
+                "1": "За период ",
+                "2": "За все время ✅",
+                "3": "рус --> англ ✅",
+                "4": "англ --> рус "
+            }
+
+            butt_dict_upd = {
+                "1": "За период ",
+                "2": "За все время ",
+                "3": "рус --> англ ",
+                "4": "англ --> рус "
+            }
+
+            create_table.insert_settings(3, 50, status, defalt_day, butt_dict, butt_dict_upd)
 
 
 async def info(message: types.Message):
